intersect: matrixFn_8006ff0c 69.5 -> 90% (f32 args drop double-precision math)

+++ include/track/intersect.h
@@ -15,7 +15,7 @@ int maybeReadDepthBuffer(int x, int y, int id);
 uint getScreenResolution(void);
 void setScreenWidth(u32 param_1);
 void clearScreenWidth(void);
-void matrixFn_8006ff0c(double param_1,double param_2,double param_3,double param_4,double param_5,
+void matrixFn_8006ff0c(f32 param_1,f32 param_2,f32 param_3,f32 param_4,f32 param_5,
                  float *param_6,short *param_7);
 void normalize(f32* x, f32* y, f32* z);
 void fn_80070234(f32* param_1);
